requirements.txt
setuptools==68.2.2
uvicorn>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain==0.0.335
langchain_core==0.1.16
langchain_openai==0.0.5
langchain_community==0.0.10
openai>=1.12.0
faiss-cpu>=1.7.4
requests==2.31.0
beautifulsoup4==4.12.2
numpy<2
tiktoken>=0.3.3
playwright==1.42.0
serpapi
google-search-results 
fastapi 
google-ads==21.3.0